pyproject.toml
[project]
name = "retail-data-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.11.4",
    "dbt-duckdb>=1.10.0",
    "fastparquet>=2025.12.0",
    "gcsfs>=2026.2.0",
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.0",
    "prefect>=3.6.16",
    "scrapy>=2.14.1",
]
